chore: upgrades, esm
