+++ test/all-features.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from 'vitest';
 
-import { getAllFeatures } from '../src/features';
-import { parseClients } from '../src/clients';
+import { getAllFeatures } from '../src/features.js';
+import { parseClients } from '../src/clients.js';
 
 describe('all features', () => {
   test('getAllFeatures()', () => {
