+++ tsconfig.json
@@ -1,27 +1,33 @@
 {
+  "$schema": "https://json.schemastore.org/tsconfig",
   "compilerOptions": {
+    "allowJs": true,
     "allowSyntheticDefaultImports": true,
     "composite": false,
     "declaration": true,
+    "declarationMap": true,
     "esModuleInterop": true,
-    "lib": ["ES2018", "ES2020", "esnext.asynciterable", "dom"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmit": true,
-    "noEmitOnError": false,
+    "forceConsistentCasingInFileNames": true,
+    "inlineSources": false,
+    "isolatedModules": true,
+    "lib": ["ES2023", "ESNext"],
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "noEmitOnError": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "outDir": "dist",
     "preserveSymlinks": true,
-    "pretty": true,
-    "removeComments": true,
+    "preserveWatchOutput": true,
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "sourceMap": true,
     "strict": true,
-    "target": "es2020",
+    "strictNullChecks": true,
+    "target": "ESNext",
     "types": ["node", "vitest/globals"],
     "useUnknownInCatchVariables": false
   },
-  "exclude": ["dist", "node_modules"],
-  "include": ["./.eslintrc.js", "./.prettierrc.js", "./tsup.config.ts", "src", "test"]
+  "exclude": ["**/dist", "**/node_modules"],
+  "include": ["src"]
 }
